+++ test/lib/stream/s3ConcurrentListObjectStream.spec.js
@@ -93,9 +93,7 @@ describe('lib/stream/s3ConcurrentListObjectStream', function () {
           );
 
           expect(nextContinuationToken).to.eql(listObjectsResponse1.NextContinuationToken);
-          expect(commonPrefixes).to.eql(_.map(listObjectsResponse1.CommonPrefixes, function (obj) {
-            return obj.Prefix;
-          }));
+          expect(commonPrefixes).to.eql(listObjectsResponse1.CommonPrefixes);
           expect(s3Objects).to.eql(listObjectsResponse1.Contents);
 
           done(error);
@@ -155,15 +153,21 @@ describe('lib/stream/s3ConcurrentListObjectStream', function () {
             Prefix: options.prefix
           });
 
-          sinon.assert.callCount(s3ConcurrentListObjectStream.push, 3);
+          sinon.assert.callCount(s3ConcurrentListObjectStream.push, 5);
 
           expect(s3ConcurrentListObjectStream.push.getCall(0).args).to.eql([
-            listObjectsResponse1.Contents[0]
+            listObjectsResponse1.CommonPrefixes[0]
           ]);
           expect(s3ConcurrentListObjectStream.push.getCall(1).args).to.eql([
-            listObjectsResponse1.Contents[1]
+            listObjectsResponse1.CommonPrefixes[1]
           ]);
           expect(s3ConcurrentListObjectStream.push.getCall(2).args).to.eql([
+            listObjectsResponse1.Contents[0]
+          ]);
+          expect(s3ConcurrentListObjectStream.push.getCall(3).args).to.eql([
+            listObjectsResponse1.Contents[1]
+          ]);
+          expect(s3ConcurrentListObjectStream.push.getCall(4).args).to.eql([
             listObjectsResponse2.Contents[0]
           ]);